+++ calendarview/src/main/java/com/haibin/calendarview/CalendarView.java
@@ -130,9 +130,9 @@ public class CalendarView extends FrameLayout {
         this.mWeekLine = findViewById(R.id.line);
         this.mWeekLine.setBackgroundColor(mDelegate.getWeekLineBackground());
         LayoutParams lineParams = (LayoutParams) this.mWeekLine.getLayoutParams();
-        lineParams.setMargins(mDelegate.getWeekLineMargin(),
-                mDelegate.getWeekBarHeight(),
-                mDelegate.getWeekLineMargin(),
+        lineParams.setMargins((int) mDelegate.getWeekLineMargin(),
+                (int) mDelegate.getWeekBarHeight(),
+                (int) mDelegate.getWeekLineMargin(),
                 0);
         this.mWeekLine.setLayoutParams(lineParams);
 
@@ -140,12 +140,11 @@ public class CalendarView extends FrameLayout {
         this.mMonthPager.mWeekPager = mWeekPager;
         this.mMonthPager.mWeekBar = mWeekBar;
         LayoutParams params = (LayoutParams) this.mMonthPager.getLayoutParams();
-        params.setMargins(0, mDelegate.getWeekBarHeight() + CalendarUtil.dipToPx(context, 1), 0, 0);
+        params.setMargins(0, (int) mDelegate.getWeekBarHeight() + CalendarUtil.dipToPx(context, 1), 0, 0);
         mWeekPager.setLayoutParams(params);
 
-
         mYearViewPager = findViewById(R.id.selectLayout);
-        mYearViewPager.setPadding(mDelegate.getYearViewPaddingLeft(), 0, mDelegate.getYearViewPaddingRight(), 0);
+        mYearViewPager.setPadding((int) mDelegate.getYearViewPaddingLeft(), 0, (int) mDelegate.getYearViewPaddingRight(), 0);
         mYearViewPager.setBackgroundColor(mDelegate.getYearViewBackground());
         mYearViewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
@@ -357,18 +356,18 @@ public class CalendarView extends FrameLayout {
                 });
 
         //颜色值过度
-        if(getBackground()!=null&& getBackground()  instanceof ColorDrawable){
-            final int intBgColor=((ColorDrawable) getBackground()).getColor();
-            final String bgColor="#"+Integer.toHexString(intBgColor);
+        if (getBackground() != null && getBackground() instanceof ColorDrawable) {
+            final int intBgColor = ((ColorDrawable) getBackground()).getColor();
+            final String bgColor = "#" + Integer.toHexString(intBgColor);
             final ValueAnimator animator = ValueAnimator.ofFloat(0.2f, 1f);
             animator.setDuration(260);
             animator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
                 @Override
                 public void onAnimationUpdate(ValueAnimator animation) {
-                    setBackgroundColor(Color.parseColor(caculateColor(bgColor,"#"+Integer.toHexString(mDelegate.getYearViewBackground()),(Float) animation.getAnimatedValue())));
+                    setBackgroundColor(Color.parseColor(caculateColor(bgColor, "#" + Integer.toHexString(mDelegate.getYearViewBackground()), (Float) animation.getAnimatedValue())));
                 }
             });
-            animator.addListener(new Animator.AnimatorListener(){
+            animator.addListener(new Animator.AnimatorListener() {
 
                 @Override
                 public void onAnimationStart(Animator animation) {
@@ -411,12 +410,13 @@ public class CalendarView extends FrameLayout {
 
     /**
      * 计算从startColor过度到endColor过程中百分比为franch时的颜色值
+     *
      * @param startColor 起始颜色 （格式#FFFFFFFF）
-     * @param endColor 结束颜色 （格式#FFFFFFFF）
-     * @param franch 百分比0.5
+     * @param endColor   结束颜色 （格式#FFFFFFFF）
+     * @param franch     百分比0.5
      * @return 返回String格式的color（格式#FFFFFFFF）
      */
-    public static String caculateColor(String startColor, String endColor, float franch){
+    public static String caculateColor(String startColor, String endColor, float franch) {
 
         int startAlpha = Integer.parseInt(startColor.substring(1, 3), 16);
         int startRed = Integer.parseInt(startColor.substring(3, 5), 16);
@@ -833,7 +833,7 @@ public class CalendarView extends FrameLayout {
      *
      * @param calendarItemHeight MonthView item height
      */
-    public final void setCalendarItemHeight(int calendarItemHeight) {
+    public final void setCalendarItemHeight(float calendarItemHeight) {
         if (mDelegate.getCalendarItemHeight() == calendarItemHeight) {
             return;
         }
